Added fn to add a course, uses fn on load
The function addCourseEntryBlock can be used in the future to add more
courses to the list
The generate function does support multicourse entry, tested my manually
calling addCourseEntryBlock.

Dynamically changing the number of courses should be as easy as adding a button to add a course.

TODO: add course button, delete course function and button